feat: multi exception folder
